Move ZetaSQL UDF tests into separate class.
